extract HalLink class
