+++ src/Context/HalModule.php
@@ -7,6 +7,7 @@
 namespace BEAR\Package\Context;
 
 use BEAR\Package\Provide\Representation\HalRenderer;
+use BEAR\Package\Provide\Representation\HalLink;
 use BEAR\Resource\RenderInterface;
 use Ray\Di\AbstractModule;
 use Ray\Di\Scope;
@@ -18,6 +19,7 @@ class HalModule extends AbstractModule
      */
     protected function configure()
     {
+        $this->bind(HalLink::class);
         $this->bind(RenderInterface::class)->to(HalRenderer::class)->in(Scope::SINGLETON);
     }
 }
